better SlashCommand implementing

+++ avalon_bot/src/avalon/config.rs
@@ -4,6 +4,7 @@ use std::mem;
 use itertools::Itertools;
 
 use discorsd::{BotState, http, UserMarkupExt};
+use discorsd::commands::Command;
 use discorsd::http::channel::{ChannelExt, embed, RichEmbed};
 use discorsd::http::ClientResult;
 use discorsd::model::commands::*;
@@ -17,9 +18,9 @@ use crate::avalon::characters::Loyalty::Evil;
 use crate::avalon::lotl::ToggleLady;
 use crate::avalon::roles::RolesCommand;
 use crate::avalon::SlashCommand;
-use crate::avalon::start::StartCommand;
-use crate::games::GameType;
 use crate::commands::addme::AddMeCommand;
+use crate::commands::start::StartCommand;
+use crate::games::GameType;
 
 #[derive(Default, Debug)]
 pub struct AvalonConfig {
@@ -30,7 +31,7 @@ pub struct AvalonConfig {
 
     /// the interaction whose message is being edited to show the game settings
     pub message: Option<Message>,
-    pub start_id: Option<CommandId>,
+    // pub start_id: Option<CommandId>,
 }
 
 impl AvalonConfig {
@@ -98,8 +99,8 @@ impl AvalonConfig {
     pub async fn update_embed(
         &mut self,
         state: &BotState<Bot>,
-        interaction: InteractionUse<Unused>,
-    ) -> http::ClientResult<InteractionUse<Used>> {
+        interaction: &InteractionUse<Used>,
+    ) -> http::ClientResult<()> {
         let embed = self.embed();
         match &mut self.message {
             Some(message) if message.channel == interaction.channel => {
@@ -119,45 +120,59 @@ impl AvalonConfig {
                 self.message = Some(new);
             }
         };
-        interaction.ack(state).await
+        Ok(())
     }
 
+    /// determine if Avalon can be started, and if it can be, include it in the list of games
+    /// available in the start command
     pub async fn start_command(
         &mut self,
         state: &BotState<Bot>,
-        commands: &mut HashMap<CommandId, Box<dyn SlashCommand<Bot>>>,
+        commands: &mut HashMap<CommandId, Box<dyn SlashCommand<Bot=Bot>>>,
         enabled: bool,
         guild: GuildId,
     ) -> ClientResult<()> {
-        let start = self.start_id
-            .and_then(|id| {
-                commands.get_mut(&id)
-                    .map(|s| s.downcast_mut::<StartCommand>().unwrap())
-                    .map(|s| (id, s))
-            });
+        let start_id = *state.bot.start.read().await.get(&guild).unwrap();
+        let start = commands.get_mut(&start_id)
+            .map(|s| s.downcast_mut::<StartCommand>().unwrap())
+            .map(|s| (start_id, s));
         match (start, enabled) {
             // update list of startable games
             (Some((id, start)), true) => {
                 if !start.0.contains(&GameType::Avalon) {
                     start.0.insert(GameType::Avalon);
+                    let Command { description, options, .. } = start.command();
                     state.client.edit_guild_command(
                         state.application_id().await,
                         guild,
                         id,
                         None,
-                        None,
-                        Some(start.command().options()),
+                        Some(description.as_ref()),
+                        Some(options),
                     ).await?;
                 }
             }
             // disable StartCommand
-            (Some((id, _)), false) => {
-                state.client.delete_guild_command(
-                    state.application_id().await,
-                    guild,
-                    id,
-                ).await?;
-                commands.remove(&id);
+            (Some((id, start)), false) => {
+                start.0.remove(&GameType::Avalon);
+                if start.0.is_empty() {
+                    state.client.delete_guild_command(
+                        state.application_id().await,
+                        guild,
+                        id,
+                    ).await?;
+                    commands.remove(&id);
+                } else {
+                    let Command { description, options, .. } = start.command();
+                    state.client.edit_guild_command(
+                        state.application_id().await,
+                        guild,
+                        id,
+                        None,
+                        Some(description.as_ref()),
+                        Some(options),
+                    ).await?;
+                }
             }
             // enable StartCommand
             (None, true) => {
@@ -167,8 +182,9 @@ impl AvalonConfig {
                     guild,
                     start.command(),
                 ).await?;
-                self.start_id = Some(command.id);
                 commands.insert(command.id, Box::new(start));
+                *state.bot.start.write().await.get_mut(&guild).unwrap() = command.id;
+                // self.start_id = Some(command.id);
             }
             // is (and should be) disabled :)
             (None, false) => {}
@@ -180,7 +196,7 @@ impl AvalonConfig {
         avalon::max_evil(self.players.len())
     }
 
-    pub fn is_setup_command(command: &dyn SlashCommand<Bot>) -> bool {
+    pub fn is_setup_command(command: &dyn SlashCommand<Bot=Bot>) -> bool {
         command.is::<StartCommand>() ||
             command.is::<AddMeCommand>() ||
             command.is::<RolesCommand>() ||
